functional refactor
